+++ packages/devtools_app/lib/src/console.dart
@@ -291,36 +291,3 @@ class DeleteControl extends StatelessWidget {
     );
   }
 }
-
-/// The type of data provider function used by the CopyToClipboard Control.
-typedef ClipboardDataProvider = String Function();
-
-/// A Console Control that copies `data` to the clipboard.
-///
-/// If it succeeds, it displays a notification with `successMessage`.
-class CopyToClipboardControl extends StatelessWidget {
-  const CopyToClipboardControl({
-    this.dataProvider,
-    this.successMessage = 'Copied to clipboard.',
-    this.tooltip = 'Copy to clipboard',
-    this.buttonKey,
-  });
-
-  final ClipboardDataProvider dataProvider;
-  final String successMessage;
-  final String tooltip;
-  final Key buttonKey;
-
-  @override
-  Widget build(BuildContext context) {
-    final disabled = dataProvider == null;
-    return ToolbarAction(
-      icon: Icons.content_copy,
-      tooltip: tooltip,
-      onPressed: disabled
-          ? null
-          : () => copyToClipboard(dataProvider(), successMessage, context),
-      key: buttonKey,
-    );
-  }
-}
