+++ packages/devtools_app/lib/src/debugger/call_stack.dart
@@ -66,8 +66,8 @@ class _CallStackState extends State<CallStack> {
         frame.frame.kind == FrameKind.kAsyncSuspensionMarker;
     final noLineInfo = frame.line == null;
 
-    final frameDescription = _descriptionFor(frame);
-    final locationDescription = _locationFor(frame);
+    final frameDescription = frame.description;
+    final locationDescription = frame.location;
 
     if (asyncMarker) {
       child = Row(
@@ -133,34 +133,5 @@ class _CallStackState extends State<CallStack> {
 
   Future<void> _onStackFrameSelected(StackFrameAndSourcePosition frame) async {
     controller.selectStackFrame(frame);
-  }
-
-  String _descriptionFor(StackFrameAndSourcePosition frame) {
-    const unoptimized = '[Unoptimized] ';
-    const none = '<none>';
-    const anonymousClosure = '<anonymous closure>';
-    const closure = '<closure>';
-    const asyncBreak = '<async break>';
-
-    if (frame.frame.kind == FrameKind.kAsyncSuspensionMarker) {
-      return asyncBreak;
-    }
-
-    var name = frame.frame.code?.name ?? none;
-    if (name.startsWith(unoptimized)) {
-      name = name.substring(unoptimized.length);
-    }
-    name = name.replaceAll(anonymousClosure, closure);
-    name = name == none ? name : '$name()';
-    return name;
-  }
-
-  String _locationFor(StackFrameAndSourcePosition frame) {
-    final uri = frame.scriptUri;
-    if (uri == null) {
-      return uri;
-    }
-    final file = uri.split('/').last;
-    return frame.line == null ? file : '$file ${frame.line}';
   }
 }
